+++ src/main.ts
@@ -1,7 +1,7 @@
 import { Linter } from "./linter";
-import type { Rule } from "eslint";
 import { fixEslint } from "./process";
 import { shouldFixOnSave } from "./shouldFixOnSave";
+import { createSuggestionCommandHandler } from "./suggestionCommand";
 
 const compositeDisposable = new CompositeDisposable();
 
@@ -36,27 +36,6 @@ function fix(
     }
 }
 
-export async function showChoicePalette<T>(
-    choices: T[],
-    // eslint-disable-next-line no-unused-vars
-    choiceToString: (choice: T) => string,
-    options?: { placeholder?: string }
-) {
-    const index = await new Promise<number | null>((resolve) =>
-        nova.workspace.showChoicePalette(
-            choices.map(choiceToString),
-            options,
-            (_, index) => {
-                resolve(index);
-            }
-        )
-    );
-    if (index == null) {
-        return null;
-    }
-    return choices[index];
-}
-
 export function activate() {
     console.log("activating...");
 
@@ -68,39 +47,7 @@ export function activate() {
     compositeDisposable.add(
         nova.commands.register(
             "apexskier.eslint.command.suggestForCursor",
-            async (editor: TextEditor) => {
-                const message = linter.getSuggestions(editor);
-
-                console.log("message", JSON.stringify(message));
-                if (!message?.fix && !message?.suggestions?.length) {
-                    nova.workspace.showWarningMessage("No suggestions found");
-                    return;
-                }
-                const choices: Array<{ title: string; fix: Rule.Fix }> = [];
-                if (message.fix) {
-                    choices.push({ title: "Fix", fix: message.fix });
-                }
-                if (message.suggestions) {
-                    choices.push(
-                        ...message.suggestions.map((suggestion) => ({
-                            title: suggestion.desc,
-                            fix: suggestion.fix,
-                        }))
-                    );
-                }
-                const choice = await showChoicePalette(
-                    choices,
-                    ({ title }) => title
-                );
-                if (choice) {
-                    editor.edit((edit) => {
-                        edit.replace(
-                            new Range(choice.fix.range[0], choice.fix.range[1]),
-                            choice.fix.text
-                        );
-                    });
-                }
-            }
+            createSuggestionCommandHandler(linter)
         )
     );
